requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
alabaster==1.0.0
# Editable install with no version control (alphanet==0.0.1.post73+ge5be1ea.d20241202)
-e /root/deep-principle/pkgs/MLFF/AlphaNet
annotated-types==0.7.0
ase==3.23.0
asttokens==3.0.0
async-timeout==4.0.3
attrs==23.2.0
autograd==1.7.0
babel==2.16.0
bcrypt==4.2.1
boltons @ file:///croot/boltons_1677628692245/work
boto3==1.34.102
botocore==1.34.102
Bottleneck @ file:///croot/bottleneck_1731058641041/work
brotlipy==0.7.0
cached_path==1.6.4
cachetools==5.5.0
cclib==1.8.1
certifi @ file:///croot/certifi_1725551672989/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
cloudpickle==3.1.0
conda @ file:///croot/conda_1694544150384/work
conda-content-trust @ file:///tmp/abs_5952f1c8-355c-4855-ad2e-538535021ba5h26t22e5/croots/recipe/conda-content-trust_1658126371814/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1685032319139/work/src
conda-package-handling @ file:///croot/conda-package-handling_1685024767917/work
conda_package_streaming @ file:///croot/conda-package-streaming_1685019673878/work
ConfigArgParse==1.7
contourpy @ file:///croot/contourpy_1732540045555/work
coverage==7.6.8
cryptography @ file:///croot/cryptography_1686613057838/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
Cython==3.0.11
dask==2024.11.2
decorator==5.1.1
Deprecated==1.2.15
dftd3==1.0.0
dftd4==3.6.0
distributed==2024.11.2
dm-tree==0.1.8
docker-pycreds==0.4.0
docutils==0.21.2
e3nn==0.4.4
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.2.0
fabric==3.2.2
filelock==3.13.4
fonttools @ file:///croot/fonttools_1713551344105/work
frozenlist==1.4.1
fsspec==2024.10.0
geometric==1.0.2
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.23.0
google-auth==2.36.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.63.0
h5py==3.11.0
huggingface-hub==0.26.3
idna @ file:///croot/idna_1666125576474/work
imagesize==1.4.1
importlib_metadata==8.5.0
iniconfig==2.0.0
invoke==2.2.0
ipdb==0.13.13
ipython==8.34.0
jedi==0.19.2
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonpatch @ file:///tmp/build/80754af9/jsonpatch_1615747632069/work
jsonpointer==2.1
kiwisolver @ file:///croot/kiwisolver_1672387140495/work
lark-parser==0.12.0
latexcodec==3.0.0
libmambapy @ file:///croot/mamba-split_1685993156657/work/libmambapy
lightning-utilities==0.11.9
lmdb==1.5.1
locket==1.0.0
-e git+ssh://git@github.com/deepprinciple/deep-principle.git@a36a3f46965f0a492368b9952d339cdddc01da90#egg=mace_torch&subdirectory=pkgs/MLFF/mace
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline @ file:///opt/conda/conda-bld/matplotlib-inline_1662014470464/work
matscipy==1.1.1
mdurl==0.1.2
monty==2024.10.21
mpmath==1.3.0
msgpack==1.1.0
multidict==6.0.5
natsort==8.4.0
networkx==3.2.1
ninja==1.11.1.1
numexpr @ file:///croot/numexpr_1730215937391/work
numpy @ file:///croot/numpy_and_numpy_base_1708638617955/work/dist/numpy-1.26.4-cp310-cp310-linux_x86_64.whl#sha256=ce136f308f566617937abac6c8f280bd61b41c12ab7eb7d8784c5795c5be477b
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py3==7.352.0
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
openpyxl==3.1.5
opt-einsum-fx==0.1.4
opt_einsum==3.4.0
packaging @ file:///croot/packaging_1678965309396/work
palettable==3.3.3
pandas @ file:///croot/pandas_1732735089971/work/dist/pandas-2.2.3-cp310-cp310-linux_x86_64.whl#sha256=3fa290e76efa7b7868273de6ab4ee4a73af177cdd88202834ad0eea6dd898d9e
paramiko==3.5.0
parso==0.8.4
partd==1.4.2
patsy==1.0.1
periodictable==1.7.1
pexpect==4.9.0
pillow @ file:///croot/pillow_1731594689713/work
platformdirs==4.3.6
plotly==5.24.1
pluggy==1.5.0
prettytable==3.12.0
prompt_toolkit==3.0.50
proto-plus==1.25.0
protobuf==5.26.1
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
py3Dmol==2.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyberny==0.6.3
pybind11==2.12.0
pybtex==0.24.0
pycosat @ file:///croot/pycosat_1666805502580/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.10.6
pydantic_core==2.27.2
pyg-lib==0.4.0+pt22cu121
Pygments==2.18.0
pyjokes==0.6.0
pymatgen==2024.11.13
PyNaCl==1.5.0
pynanoflann @ git+https://github.com/u1234x1234/pynanoflann.git@aea45ba149dba7336c921c6d685ec14aca57deb2
pyOpenSSL @ file:///croot/pyopenssl_1677607685877/work
pyparsing @ file:///opt/conda/conda-bld/pyparsing_1661452539315/work
pyprof==1.0.0
pyscf==2.6.2
pyscf-dispersion==1.0.2
pyscf-doci==0.1.0
pyscf-forge==1.0.0
pyscf-properties==0.1.0
pyscf-qsdopt==0.1.0
pyscf-semiempirical==0.1.1
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
pytest==8.3.3
pytest-cov==6.0.0
python-dateutil @ file:///croot/python-dateutil_1716495738603/work
python-hostlist==2.0.0
pytorch-lightning==2.4.0
pytz @ file:///croot/pytz_1713974312559/work
PyYAML==6.0.2
rdkit==2024.3.6
regex==2024.5.10
requests==2.32.3
rich==13.9.4
rmsd==1.5.1
rsa==4.9
ruamel.yaml @ file:///croot/ruamel.yaml_1666304550667/work
ruamel.yaml.clib @ file:///croot/ruamel.yaml.clib_1666302247304/work
s3transfer==0.10.1
safetensors==0.5.3
scikit-learn==1.3.0
scipy==1.13.1
sentry-sdk==2.22.0
setproctitle==1.3.5
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
spglib==2.5.0
Sphinx==8.1.3
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
stack-data==0.6.3
statsmodels==0.14.4
sympy==1.13.3
tabulate==0.9.0
tblib==3.0.0
tenacity==9.0.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
timm==1.0.15
tomli==2.2.1
toolz @ file:///croot/toolz_1667464077321/work
torch==2.2.1+cu121
torch-ema==0.3
torch-geometric==2.6.1
torch-scatter==2.1.2
torch_cluster==1.6.3+pt22cu121
torch_sparse==0.6.18+pt22cu121
torch_spline_conv==1.2.2+pt22cu121
torchaudio==2.2.1+cu121
torchdiffeq==0.2.5
torchmetrics==1.6.0
torchvision==0.17.1+cu121
tornado==6.4.2
tqdm==4.67.1
traitlets @ file:///croot/traitlets_1718227057033/work
triton==2.2.0
typing_extensions==4.12.2
tzdata @ file:///croot/python-tzdata_1690578112552/work
uncertainties==3.2.2
unicodedata2 @ file:///croot/unicodedata2_1713212950228/work
urllib3 @ file:///croot/urllib3_1686163155763/work
wandb==0.19.8
wcwidth==0.2.13
Werkzeug==3.0.3
wrapt==1.17.0
xgboost==2.0.3
xtb @ file:///home/conda/feedstock_root/build_artifacts/xtb-python_1732695631550/work/xtb-22.1-cp310-cp310-linux_x86_64.whl#sha256=3d01ba90c8ac8d6fd08ac5bcf241506cdef181949bde4e4487d58c2565d424a4
yarl==1.9.4

zict==3.0.0
zipp==3.21.0
zstandard @ file:///croot/zstandard_1677013143055/work
